Move ApiKey action and jspx to better organize files, update paths so links and buttons still work
