+++ onebusaway-admin-webapp/src/main/java/org/onebusaway/webapp/actions/admin/apikeymanagement/ApiKeyAction.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.onebusaway.webapp.actions.admin;
+package org.onebusaway.webapp.actions.admin.apikeymanagement;
 
 import java.util.List;
